+++ src/ConstructNamedParameters/Contracts/CreatableFromArrayInterface.php
@@ -1,0 +1,14 @@
+<?php
+namespace ConstructNamedParameters\Contracts;
+
+interface CreatableFromArrayInterface
+{
+    /**
+     * Create an instance of the object based on the key and values from the array
+     *
+     * @param array $values
+     *
+     * @return $this
+     */
+    public static function create(array $values);
+}