+++ src/ConstructNamedParameters/Contracts/ExportableToArrayInterface.php
@@ -1,0 +1,12 @@
+<?php
+namespace ConstructNamedParameters\Contracts;
+
+interface ExportableToArrayInterface
+{
+    /**
+     * Return an array with properties as key values
+     *
+     * @return array
+     */
+    public function toArray();
+}